+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Semantic versioning 2.0 is followed.
 
+# 1.1.0
+
+- PHP 8.2 support
+- Use php://temp stream wrapper for storing responses
+- Response headers were not parsed correctly
+
 # 1.0.1
 
 - Streamed response body and headers were not cleared for every request